Add CompanyModel class with navigation properties
Added a new `CompanyModel` class to the `JobUpdatesAPI.Models` namespace.
This class includes properties for `CompanyId`, `Name`, `Jobs`, `Job`,
and `Agency`. The `CompanyId` is marked as the primary key using the
`[Key]` attribute. Navigation properties were added to represent
relationships with jobs and parent companies. Also, an override for
the `ToString` method was implemented to provide a string
representation of the model.